projects/requirements.txt
requests>=2.31.0
tenacity>=8.2.2
tqdm>=4.66.0
pymupdf>=1.23.0
faiss-cpu>=1.7.4
openai>=1.12.0
python-dotenv>=1.0.0
google-generativeai